prospect-search/linkedin/requirements.txt
requests>=2.31
python-dotenv>=1.0
thefuzz>=0.22
